+++ src/main/java/com/mivek/parser/TAFParser.java
@@ -7,7 +7,6 @@ import com.mivek.model.BeginningValidity;
 import com.mivek.model.Cloud;
 import com.mivek.model.PROBChange;
 import com.mivek.model.TAF;
-import com.mivek.model.TEMPOChange;
 import com.mivek.model.TemperatureDated;
 import com.mivek.model.Validity;
 import com.mivek.model.Visibility;
@@ -15,6 +14,7 @@ import com.mivek.model.WeatherCondition;
 import com.mivek.model.trend.AbstractWeatherChange;
 import com.mivek.model.trend.BECMGChange;
 import com.mivek.model.trend.FMChange;
+import com.mivek.model.trend.TEMPOChange;
 import com.mivek.utils.Converter;
 import com.mivek.utils.Regex;
